refactor: 예외 import 처리
